+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/GamesServiceClient.java
@@ -262,4 +262,8 @@ public class GamesServiceClient extends VPinStudioClientService {
   public boolean resetHighscore(int gameId) {
     return getRestClient().delete(API + "games/reset/" + gameId);
   }
+
+  public GameScoreValidation getGameScoreValidation(int gameId) {
+    return getRestClient().get(API + "games/scorevalidation/" + gameId, GameScoreValidation.class);
+  }
 }