+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameValidationService.java
@@ -502,7 +502,7 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     GameScoreValidation validation = new GameScoreValidation();
     validation.setValidScoreConfiguration(true);
 
-    boolean played = tableDetails.getNumberPlays() > 0;
+    boolean played = tableDetails.getNumberPlays() != null && tableDetails.getNumberPlays() > 0;
     ScoringDB scoringDB = systemService.getScoringDatabase();
     List<String> vpRegEntries = highscoreService.getVPRegEntries();
     List<String> highscoreFiles = highscoreService.getHighscoreFiles();
@@ -535,16 +535,25 @@ public class GameValidationService implements InitializingBean, PreferenceChange
       return validation;
     }
 
-    //no fields are set
-    if (StringUtils.isEmpty(rom) && StringUtils.isEmpty(tableName) && StringUtils.isEmpty(hsName)) {
-      validation.setRomIcon(GameScoreValidation.ERROR_ICON);
-      validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
-      validation.setRomStatus(GameScoreValidation.STATUS_FIELDS_NOT_SET);
+    //not played and no highscore file found
+    if (!played && !StringUtils.isEmpty(hsName) && !highscoreFiles.contains(hsName)) {
+      validation.setHighscoreFilenameIcon(GameScoreValidation.UNPLAYED_ICON);
+      validation.setHighscoreFilenameIconColor(GameScoreValidation.OK_COLOR);
+      validation.setHighscoreFilenameStatus(GameScoreValidation.STATUS_NOT_PLAYED_HSFILE_NOT_FOUND);
+      return validation;
+    }
+
+    //not played and the ROM VPReg.stg entry not found
+    if (!played && !vpRegEntries.contains(String.valueOf(rom).toLowerCase()) && !vpRegEntries.contains(rom) && !game.getNvRamFile().exists()) {
+      validation.setRomIcon(GameScoreValidation.UNPLAYED_ICON);
+      validation.setRomIconColor(GameScoreValidation.OK_COLOR);
+      validation.setRomStatus(GameScoreValidation.STATUS_NOT_PLAYED_NO_MATCH_FOUND);
       return validation;
     }
 
     //no fields are set
     if (StringUtils.isEmpty(rom) && StringUtils.isEmpty(tableName) && StringUtils.isEmpty(hsName)) {
+      validation.setValidScoreConfiguration(false);
       validation.setRomIcon(GameScoreValidation.ERROR_ICON);
       validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
       validation.setRomStatus(GameScoreValidation.STATUS_FIELDS_NOT_SET);
@@ -552,7 +561,8 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     }
 
     //ROM is not supported
-    if (!StringUtils.isEmpty(rom) && scoringDB.isNotSupported(rom)) {
+    if (!StringUtils.isEmpty(rom) && (scoringDB.getNotSupported().contains(rom) ||!scoringDB.getSupportedNvRams().contains(rom))) {
+      validation.setValidScoreConfiguration(false);
       validation.setRomIcon(GameScoreValidation.ERROR_ICON);
       validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
       validation.setRomStatus(GameScoreValidation.STATUS_ROM_NOT_SUPPORTED);
@@ -560,10 +570,29 @@ public class GameValidationService implements InitializingBean, PreferenceChange
     }
 
     //Highscore file is not supported
-    if (!StringUtils.isEmpty(hsName) && scoringDB.isNotSupported(hsName)) {
+    if (!StringUtils.isEmpty(hsName) && scoringDB.getNotSupported().contains(hsName)) {
+      validation.setValidScoreConfiguration(false);
       validation.setRomIcon(GameScoreValidation.ERROR_ICON);
       validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
       validation.setRomStatus(GameScoreValidation.STATUS_HSFILE_NOT_SUPPORTED);
+      return validation;
+    }
+
+    //game has been played, but the text file has not been generated
+    if (played && !StringUtils.isEmpty(hsName) && !highscoreFiles.contains(hsName)) {
+      validation.setValidScoreConfiguration(false);
+      validation.setRomIcon(GameScoreValidation.ERROR_ICON);
+      validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
+      validation.setRomStatus(GameScoreValidation.STATUS_PLAYED_HSFILE_NOT_FOUND);
+      return validation;
+    }
+
+    //game has been played, but the .nvram or VPReg has not been found
+    if (played && !StringUtils.isEmpty(rom) && !vpRegEntries.contains(rom) && !vpRegEntries.contains(tableName) && !game.getNvRamFile().exists()) {
+      validation.setValidScoreConfiguration(false);
+      validation.setRomIcon(GameScoreValidation.ERROR_ICON);
+      validation.setRomIconColor(GameScoreValidation.ERROR_COLOR);
+      validation.setRomStatus(GameScoreValidation.STATUS_PLAYED_NO_MATCH_FOUND);
       return validation;
     }