+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/TextEditorController.java
@@ -1,6 +1,7 @@
 package de.mephisto.vpin.ui;
 
 import de.mephisto.vpin.commons.fx.DialogController;
+import de.mephisto.vpin.commons.utils.FileUtils;
 import de.mephisto.vpin.commons.utils.WidgetFactory;
 import de.mephisto.vpin.restclient.textedit.TextFile;
 import de.mephisto.vpin.restclient.textedit.VPinFile;
@@ -9,6 +10,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
@@ -21,6 +23,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.URL;
+import java.text.DateFormat;
+import java.util.Date;
 import java.util.ResourceBundle;
 
 import static de.mephisto.vpin.ui.Studio.client;
@@ -46,8 +50,16 @@ public class TextEditorController implements Initializable, DialogController {
   @FXML
   private Button closeBtn;
 
+  @FXML
+  private Label size;
+
+  @FXML
+  private Label lastModified;
+
   private RichText richText;
   private VPinFile file;
+
+  private boolean saved = false;
 
   @FXML
   private void onClose() {
@@ -58,12 +70,15 @@ public class TextEditorController implements Initializable, DialogController {
   private void onSave(ActionEvent e) {
     Stage stage = (Stage) ((Button) e.getSource()).getScene().getWindow();
 
+    saved = true;
     this.saveBtn.setDisable(true);
     this.saveAndCloseBtn.setDisable(true);
     this.closeBtn.setDisable(true);
 
     try {
-      client.getTextEditorService().save(file, this.richText.getCodeArea().getText());
+      TextFile save = client.getTextEditorService().save(file, this.richText.getCodeArea().getText());
+      lastModified.setText(DateFormat.getDateTimeInstance().format(save.getLastModified()));
+      size.setText(FileUtils.readableFileSize(save.getSize()));
     } catch (Exception ex) {
       WidgetFactory.showAlert(stage, "Error", "Error saving file: " + ex.getMessage());
     }
@@ -115,6 +130,9 @@ public class TextEditorController implements Initializable, DialogController {
   public void load(VPinFile file) {
     this.file = file;
     TextFile value = client.getTextEditorService().getText(file);
+    lastModified.setText(DateFormat.getDateTimeInstance().format(value.getLastModified()));
+    size.setText(FileUtils.readableFileSize(value.getSize()));
+
     richText = new RichText(value.getContent());
 
     VirtualizedScrollPane scrollPane = new VirtualizedScrollPane(richText.getCodeArea());
@@ -128,5 +146,9 @@ public class TextEditorController implements Initializable, DialogController {
   @Override
   public void onDialogCancel() {
 
+  }
+
+  public boolean isSaved() {
+    return saved;
   }
 }